Started libgui implementation

+++ src/userspace/libc/stdint.h
@@ -12,7 +12,7 @@ typedef unsigned short u16;
 typedef signed int s32;
 typedef unsigned int u32;
 
-typedef signed long s64;
-typedef unsigned long u64;
+typedef signed long long s64;
+typedef unsigned long long u64;
 
 #endif
